Avoid NPE during tab switches by
This exception was being made when changing tabs:

        java.lang.NullPointerException
                at android.app.BackStackRecord.run(BackStackRecord.java:606)
                at android.app.FragmentManagerImpl.execPendingActions(FragmentMa
                at android.app.FragmentManagerImpl$1.run(FragmentManager.java:42
                at android.os.Handler.handleCallback(Handler.java:605)
                at android.os.Handler.dispatchMessage(Handler.java:92)
                at android.os.Looper.loop(Looper.java:137)

I'm not 100% sure why, but it seems that you can't just wire the
ActionBar.TabListener callbacks straight to your activity and be done with
it.  This FragmentTabListener class is required.  This one is taken from
Codepath with minor modifications to fit this Android 14+ only codebase.