Yak Shaving for Stream View: layouts, Stream and StockImage objects
I'll do the "featured" stream in a new activity, with a background ListView
which holds the image for each activity (to be loaded from StockImage).
The activity will intercept gestures and move the ListView during touch,
smoothscroll to the nearest one on touch up.  Then an onScroll can
"animate" the fading in of the description, background gradient, etc.
The bottom right-hand button will switch in either the donate or RSVP
fragment, depending on what is being looked at.