Fine-tuning the smooth scrolling and fading-in labels
I tried to make it so that the labels hide themselves when a scroll over an
item is in progress.  Also, there is a loop in here which measures the
actual velocity and acceleration of the view as it is being drawn.  This
took some experimenting, so this code is a bit messy.